airflow_dbt_metabase_lab/requirements.txt
pandas
pyarrow
sqlalchemy
psycopg2-binary
apache-airflow-providers-postgres
dbt-postgres